requirements (1).txt
google-genai>=1.46.0
pydantic>=2.11.7
elevenlabs>=2.20.1
python-dotenv>=1.1.0
flask>=3.0.0
flask-cors>=4.0.0
moviepy>=2.2.1
opencv-python>=4.12.0.88
